# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "ДКП API",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/v1/ocr/passport": {
            "post": {
                "tags": [
                    "Распознавание"
                ],
                "summary": "Распознать паспорт",
                "operationId": "749f6bd3d3541af5439c27b1fbcac1bb",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "file"
                                ],
                                "properties": {
                                    "file": {
                                        "description": "Изображение паспорта (jpeg, jpg, png, pdf, max 20MB)",
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Паспорт успешно распознан",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Passport recognized successfully"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Person"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Ошибка распознавания или валидации",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Failed to recognize passport"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "cookieAuth": []
                    }
                ]
            }
        },
        "/v1/ocr/sts": {
            "post": {
                "tags": [
                    "Распознавание"
                ],
                "summary": "Распознать СТС",
                "operationId": "fd08663db59d3e481dd50b8f987400cb",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "file"
                                ],
                                "properties": {
                                    "file": {
                                        "description": "Изображение СТС (jpeg, jpg, png, pdf, max 20MB)",
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "СТС успешно распознан",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "STS recognized successfully"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Car"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Ошибка распознавания или валидации",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Failed to recognize STS"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "cookieAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Person": {
                "title": "Person",
                "description": "Физическое лицо (продавец или покупатель)",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "surname": {
                        "type": "string",
                        "example": "Иванов"
                    },
                    "name": {
                        "type": "string",
                        "example": "Иван"
                    },
                    "fathername": {
                        "type": "string",
                        "example": "Иванович",
                        "nullable": true
                    },
                    "birthdate": {
                        "type": "string",
                        "example": "01.01.1990"
                    },
                    "country": {
                        "type": "string",
                        "example": "Россия"
                    },
                    "index": {
                        "type": "string",
                        "example": "123456"
                    },
                    "region": {
                        "type": "string",
                        "example": "Москва"
                    },
                    "passport": {
                        "$ref": "#/components/schemas/Passport"
                    },
                    "avatar": {
                        "type": "string",
                        "example": "https://example.com/avatar.jpg",
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "Passport": {
                "title": "Passport",
                "description": "Паспортные данные",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "serie": {
                        "type": "string",
                        "example": "1234"
                    },
                    "number": {
                        "type": "string",
                        "example": "567890"
                    },
                    "issuer": {
                        "type": "string",
                        "example": "ОУФМС России по г. Москве"
                    },
                    "issue_date": {
                        "type": "string",
                        "example": "01.01.2010"
                    }
                },
                "type": "object"
            },
            "Car": {
                "title": "Car",
                "description": "Автомобиль",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "vin": {
                        "type": "string",
                        "example": "WVWZZZ1JZXW123456"
                    },
                    "sts": {
                        "type": "string",
                        "example": "77AA123456"
                    },
                    "pts": {
                        "type": "string",
                        "example": "77ББ123456"
                    },
                    "plates": {
                        "type": "string",
                        "example": "А123БВ777"
                    },
                    "model": {
                        "type": "string",
                        "example": "Toyota Camry"
                    },
                    "type_category": {
                        "type": "string",
                        "example": "Легковой автомобиль"
                    },
                    "issue_year": {
                        "type": "string",
                        "example": "2020"
                    },
                    "engine_model": {
                        "type": "string",
                        "example": "2AR-FE"
                    },
                    "engine_number": {
                        "type": "string",
                        "example": "1234567"
                    },
                    "chassis_number": {
                        "type": "string",
                        "example": "12345678901234567",
                        "nullable": true
                    },
                    "body_number": {
                        "type": "string",
                        "example": "12345678901234567"
                    },
                    "color": {
                        "type": "string",
                        "example": "Черный"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "cookieAuth": {
                "type": "apiKey",
                "name": "laravel_session",
                "in": "cookie"
            }
        }
    },
...
}